+++ packages/fluentui/react-northstar/src/themes/teams-forced-colors/components/Dropdown/dropdownItemStyles.ts
@@ -1,0 +1,9 @@
+import { ComponentSlotStylesPrepared } from '@fluentui/styles';
+import { DropdownItemStylesProps } from '../../../teams/components/Dropdown/dropdownItemStyles';
+import { DropdownVariables } from '../../../teams/components/Dropdown/dropdownVariables';
+
+export const dropdownItemStyles: ComponentSlotStylesPrepared<DropdownItemStylesProps, DropdownVariables> = {
+  root: () => ({
+    border: 'none',
+  }),
+};